refactor(slice-45): mirror tests and split slice16 mega-suite
Align the unit tree with server/cli packages and replace the 2.3k-line
Slice 16 suite with focused pipeline modules plus shared helpers so
gate ignore paths and nested repo-root discovery stay correct.
