+++ .github/workflows/ci.yml
@@ -172,9 +172,9 @@ jobs:
         run: |
           pytest --tb=short -q \
             --ignore=tests/contract \
-            --ignore=tests/test_postgres_store_integration.py \
-            --ignore=tests/test_postgres_dense_retrieval.py \
-            --ignore=tests/test_postgres_sparse_hybrid.py \
+            --ignore=tests/server/db/test_postgres_store_integration.py \
+            --ignore=tests/server/db/test_postgres_dense_retrieval.py \
+            --ignore=tests/server/db/test_postgres_sparse_hybrid.py \
             -m "not integration" \
             --cov=server.core.guards.search_index_plan \
             --cov=server.core.guards.search_index_guard \
@@ -236,9 +236,9 @@ jobs:
         # Contract suite is parametrized (mongo+postgres); the mongo param skips
         # here because only Postgres is provisioned.
         run: |
-          pytest tests/test_postgres_store_integration.py \
-                 tests/test_postgres_dense_retrieval.py \
-                 tests/test_postgres_sparse_hybrid.py \
+          pytest tests/server/db/test_postgres_store_integration.py \
+                 tests/server/db/test_postgres_dense_retrieval.py \
+                 tests/server/db/test_postgres_sparse_hybrid.py \
                  tests/contract/test_storage_backend_contract.py \
                  --tb=short -q -rs \
                  --cov=server.core.retrieval.retriever_postgres \
