+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Slice 45 (rerank package)** — moved `reranker` + `local_reranker` into `server/core/rerank/`; shim aliases keep old import/patch paths working.
 - **Slice 45 (db theme packages)** — moved ports/factory into `server/db/ports/`, Mongo adapters into `server/db/mongo/`, Postgres adapters + `schema.sql` into `server/db/postgres/`; shim aliases at old `server.db.*` paths.
 - **Slice 45 (fat-surface slim)** — extracted Bayesian/stale-status helpers to `server/api/experiments_lifecycle.py` (`experiments.py` 553→388) and CLI watch/summary presentation to `cli/display.py` (`cli/main.py` 519→246).
+- **Slice 45 (tests mirror + mega-suite split)** — flat `tests/test_*.py` mirrored under `tests/server/{core,db,api,models}/`, `tests/cli/`, `tests/scripts/`; `test_slice16_parallel_sweep.py` (~2296) split into focused `tests/server/core/pipeline/` modules + `tests/helpers/pipeline_sweep.py`.
 - **Nightly Meterian OSS path** — `nightly.yml` Meterian job uses `oss: true` (no `METERIAN_API_TOKEN`), pins scanners to Python + Node (`cli_args`), and archives `meterian-<run>` artifacts (HTML, JUnit, SARIF, CycloneDX + CSV SBOM) for comparison with Anchore (**IMPLEMENTED**; not yet observed on a green nightly run).
 - **Shared coverage floors 95/90/95/95** (DECISIONS #142) — FE Vitest **95/90/95/95**; BE **95/90/n/a/95** via `fail_under=95` + `scripts/check_backend_coverage_floors.py` (stmts/br/lines from coverage JSON; functions n/a on coverage.py). Gap tests raised TOTAL ≈97.7%.
 - **Fair coverage metric floors** (DECISIONS #141) — FE **95/90/95/95**; BE interim policy 92/85 with `fail_under=90` — **superseded for BE by #142**.
@@ -62,7 +63,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **`STORAGE_BACKEND=mongodb` canonical token** — default and health/docs/CI use `mongodb` (matches YAML `database_provider: mongodb`); legacy `STORAGE_BACKEND=mongo` still accepted and normalized. Operator guides (postgres-setup, configuration, troubleshooting, CLI health examples) updated.
 - **Postgres/Supabase operator docs parity** — `postgres-setup.md` mirrors Mongo/SIE scaffolding (env tables, Path B, diagnostics); Supabase clarified as hosted Postgres under `STORAGE_BACKEND=postgres` + canonical `DATABASE_URL` (optional `SUPABASE_URI` alias added in Slice 37); empty-state dashboard offers both `configs/mongodb/` and `configs/supabase/` examples.
 - **CLI `indexes` on Postgres** — `indexes list|reset` exits not-applicable when `STORAGE_BACKEND` is not mongodb (no Atlas/`MONGODB_URI` surprise).
-- **Unit vs live test split** — `quality-gates.sh` / CI backend / `pre-push-gates.sh` exclude live Mongo/Postgres suites (`tests/contract/`, `tests/test_postgres_*.py`, `-m "not integration"`); live coverage stays in `postgres-integration` and `mongo-integration` jobs. Session-scoped Postgres pool + schema advisory lock avoid DDL deadlocks when live suites do run interleaved.
+- **Unit vs live test split** — `quality-gates.sh` / CI backend / `pre-push-gates.sh` exclude live Mongo/Postgres suites (`tests/contract/`, `tests/server/db/test_postgres_*.py`, `-m "not integration"`); live coverage stays in `postgres-integration` and `mongo-integration` jobs. Session-scoped Postgres pool + schema advisory lock avoid DDL deadlocks when live suites do run interleaved.
 - **MongoDB/Postgres boundary hygiene** — rename agnostic `mongo_*` API helpers to port verbs; `retriever.py` → `retriever_mongo.py`; MongoDB-only docstring markers; settings `ensure_storage_ready()`; ops/UI copy no longer assumes Atlas on `--postgres`
 - **Example configs split by backend** — YAML examples live under `configs/mongodb/` and `configs/supabase/` with mirrored stems (`example-local.yaml`, `example-voyage.yaml`, `example-sie.yaml`, parallel/bayesian/unified variants). Shared queries remain at `configs/questions.example.json`. Supabase configs use `database_provider: supabase`; dense, sparse, and hybrid retrieval are available on both backends (Slice 35).
 - **Slice 39 — Demo-ready dashboard polish** — result-led experiment cards, lifecycle-first detail hierarchy, shared responsive visual tokens, visible focus treatment, WCAG AA contrast, and reduced-motion support; API calls, controls, routing, and polling contracts remain unchanged
