+++ docs/plan/DECISIONS.md
@@ -157,3 +157,4 @@
 | 151 | 2026-07-28 | 45 | Land `server/core/rerank/` (`reranker`, `local_reranker`); `sys.modules` shims at old paths | Completes core Feature provider cluster from theme map; search imports canonical `server.core.rerank.reranker` | Fold into #150 / leave flat |
 | 152 | 2026-07-28 | 45 | Land `server/db/{ports,mongo,postgres}/`; shims at old flat `server.db.*` paths except **no** `server/db/postgres.py` (collides with package); pool API re-exported on `server.db.postgres` package; `schema.sql` co-located with `postgres.py` | Must hotspot #2; preserves Protocols + factory; `from server.db import postgres` / `close_pool` stays valid | Split db across multiple PRs / invent `pg/` package rename |
 | 153 | 2026-07-28 | 45 | Fat-surface slim: `experiments_lifecycle.py` + `cli/display.py`; keep routers/commands as façades | Execution order step 3 after db; pre→post lines experiments 553→388, cli/main 519→246 | Defer slim until FE primitives / leave god modules |
+| 154 | 2026-07-28 | 45 | Mirror flat tests under `tests/server/`, `tests/cli/`, `tests/scripts/`; split Slice 16 mega-suite into `tests/server/core/pipeline/` + `helpers/pipeline_sweep.py`; update gate/CI postgres ignore paths | Step 4 after import paths stable; keep `contract/` + root `conftest.py` | Mirror without mega-suite split / invent `tests/unit/` only |
