+++ tests/cli/test_cli_print_summary.py
