+++ tests/cli/test_indexes_cmd.py
