+++ tests/server/api/test_sweep_endpoint.py
