+++ tests/server/core/embedding/test_embedder_factory.py
