+++ tests/server/core/embedding/test_local_embedder.py
