+++ tests/server/core/embedding/test_sie_embedder.py
