+++ tests/server/core/embedding/test_sie_embedder_cancel.py
