+++ tests/server/core/embedding/test_sie_settings.py
