+++ tests/server/core/guards/test_search_index_guard.py
