+++ tests/server/core/guards/test_search_index_plan.py
