+++ tests/server/core/guards/test_sie_guard.py
