+++ tests/server/db/test_storage_mode.py
