+++ tests/server/db/test_store_factory.py
