+++ tests/server/db/test_supabase_uri_alias.py
